+++ CHANGELOG.rst
@@ -1,15 +1,22 @@
+0.1.3b (21.04.2012)
+*******************
+
+* Fixed S3Storage._save
+* Fixed S3Storage.__init__
+* Fixed upload shortcut
+
 0.1.2b (21.04.2012)
--------------------
+*******************
 
 * Fixed S3Storage._save
 
 0.1.1b (18.04.2012)
--------------------
+*******************
 
 * Fixed S3Storage.listdir
 * Fixed README.
 
 0.1b (18.04.2012)
------------------
+*****************
 
 * S3 backend for django file storage